Signed-off-by: suod-NithishKarthik <nithishkarthik01@gmail.com> fix[UI]: Remove 'keepMounted' prop on 'Popper' component
In the User component, we have the Popper component, for which, the keepMounted prop has been set to True, which has unfortunately mounted it to the top left corner where our 'Meshery Logo' ListItem is present.
So, because of the zIndex, it blocks the Click action.

Fix

set keepMounted to False

The keepMounted prop, I put it there only because that would have a marginal impact on responsiveness, and if this is the cost, then we can remove it.

Resolves Issue #3180