remove "meshery-operator" on shutdown
when meshery server is terminated, it'll remove the meshery-operator
from your cluster. It will actually only uninstall the meshery-operator
helm chart which takes with these objects with itself:
meshery-controller-role
meshery-operator-role
meshery-proxy-role
meshery-metrics-reader
meshery-controller-rolebinding
meshery-operator-rolebinding
meshery-proxy-rolebinding
In in-cluster deployments, mesheryctl will do this and when run locally,
the code added with this commit will.

Signed-off-by: Rudraksh Pareek <rudrakshpareek3601@gmail.com>