mesheryctl: config: update error message when connection refused
Update the error message displayed to the user if the connection is
refused by Meshery Server, when running the command `mesheryctl system
config aks|eks|gke|minikube`.
The method setToken() has been moved and is used before communicating
with a cluster provider, to check beforehand whether Meshery
Server is running or not.

Signed-off-by: Luigi Antonio Junior De Matteis <luigi.dematteis.linuxdev@gmail.com>