Fixed email-on-failure job in stable release
This job was added to notify support@layer5 via an email if any of the
jobs defined in the build-and-relase-stable workflow failed.

Problem: The subject and the body of the email being sent were
incorrect.

Reason: Previously the subject and the body were being generated using
the variables assigned to the current job in execution i.e.
email-on-failure.

Solution: Now the failed jobs are determined by iterating over all the
jobs in the needs array and checking their result/status. If the status
is `failure` then that particular job is put in the failed_jobs
variable. And its details are appended in the failed_details variable.
Now when the email will be sent, it will contain the right details.

Signed-off-by: VaibhavMalik4187 <vaibhavmalik2018@gmail.com>
Signed-off-by: Dhananjay Sahoo <kanha94377@gmail.com>